[DEV] Add reserved entry in sysmon + modify filter
Settings to be send to sysmon plugin should be formated as:
monitor=proc1|proc2;module=0xfffff, where monitor part
represents process to monitor with sysmon plugin and
module part is a bit field used to activate given
modules. In order to keep sysmon as it was before,
we should put module to 0xfffffff7f and monitor to old
sysmon setting string.